+++ Assets/Scripts/Game/ScoreManager.cs
@@ -1,24 +1,55 @@
 ﻿using UnityEngine;
+using System;
 using System.Collections;
+using System.Collections.Generic;
 using UnityEngine.UI;
+
+public class Score : IComparable<Score>
+{
+    public string name;
+    public int score;
+
+    public Score(string _name, int _score)
+    {
+        name = _name;
+        score = _score;
+    }
+
+    public int CompareTo(Score s)
+    {
+        return s.score.CompareTo(score);
+    }
+}
 
 public class ScoreManager : MonoBehaviour {
 
     public Text scoreText;
 
     public void Refresh() {
-        scoreText.text = "";
-        if (PlayerPrefs.GetInt("Player1") != 0) {
-            scoreText.text += "Stranger "  + PlayerPrefs.GetInt("Stranger") + "\n";
-        }
+
+        List<Score> scores =  new List<Score>();
+
         if (PlayerPrefs.GetInt("Player2") != 0) {
-            scoreText.text += "Alien Bear "  + PlayerPrefs.GetInt("Alien Bear") + "\n";
-        }
-        if (PlayerPrefs.GetInt("Player3") != 0) {
-            scoreText.text += "Scrap "  + PlayerPrefs.GetInt("Scrap") + "\n";
+            scores.Add (new Score("Alien Bear", PlayerPrefs.GetInt ("Alien Bear")));
         }
         if (PlayerPrefs.GetInt("Player4") != 0) {
-            scoreText.text += "Hunter "  + PlayerPrefs.GetInt("Hunter") + "\n";
+            scores.Add (new Score("Hunter", PlayerPrefs.GetInt ("Hunter")));
         }
+        if (PlayerPrefs.GetInt("Player3") != 0) {
+            scores.Add (new Score("Scrap", PlayerPrefs.GetInt ("Scrap")));
+        }
+        if (PlayerPrefs.GetInt("Player1") != 0) {
+            scores.Add (new Score("Stranger", PlayerPrefs.GetInt ("Stranger")));
+        }
+
+        scores.Sort ();
+
+        string textScores = "";
+        foreach( Score score in scores)
+        {
+            textScores += score.name + " " + score.score + "\n";
+        }
+
+        scoreText.text = textScores;
     }
 }